Add files via upload
A little bit of code I researched and put together which runs only in Windows. I'm not sure which file is the final code, but the idea was that this DUBIOUS little program could be run from a USB. The user inputs what they are looking for and the program searches whatever computer it's on and either copies the files their input turns up in (or lists them). I know how morally dubious this code is and I take no responsibility for it being copied and used for bad things. I simply made it to see if I could do it. I'm posting the code, so it is out of my hands. Also however, some of this may have been copied from other sources/people who I did not keep a record of at the time so if you recognise any of it (aside from the general OS code) please post the source for me as I don't intend to steal it :-)